1. Took IMietgegenstand out of ILandfahrzeug and added into the classes directly instead.
2. Removed unnecessary icon links from the html page. iconify keeps catching errors about unidentified objects? WIP

3. Continued on creating the necessary tables in the database. Currently having a question/ problem about how to save images in the database.

4. Wanted to make an encapsulated image class that i use for managing all my images in this and future projects. Still WIP and I'll most likely keep it for the end of the project.

+++ CoreWebAPI/VRentalsRESTAPI/VRentalsClasses/Interfaces/IMietgegenstand.cs
@@ -25,7 +25,7 @@ namespace VRentalsClasses.Interfaces
         public List<Schaden>? SchadenListe { get; set; }
 
         // Für welchen Preis wird es vermietet?
-        public int? MietPreis { get; set; }
+        public double? MietPreis { get; set; }
         // Was ist der jetzige Zustand des Gegenstands? Ist es vermietbar? Wird es gerade gewartet? etc... 
         public GegenstandZustandTyp GegenstandZustand { get; set; }
         // Gehört der Gegenstand einer Kategorie an?
